feat(Plugin): Add plugin
Plugin: mass renaming
