+++ MiAZ/data/docs/VERSION
@@ -1,1 +1,1 @@
-0.0.4+build.304
+0.0.4+build.313
