+++ MiAZ/data/plugins/massrename.plugin
@@ -1,0 +1,10 @@
+[Plugin]
+Module=massrename
+Name=MiAZMassRename
+Loader=Python3
+Description=Mass renaming of items
+Authors=Tomás Vírseda <tomasvirseda@gmail.com>
+Copyright=Copyright © 2023 Tomás Vírseda
+Website=http://github.com/t00m/MiAZ
+Help=http://github.com/t00m/MiAZ/README.adoc
+Version=0.1
